# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个 Go 项目的 API 文档。",
        "title": "Trae Go API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/books": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取所有书籍信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "获取书籍列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Book"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "添加一本新书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "创建书籍",
                "parameters": [
                    {
                        "description": "书籍信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据 ID 获取书籍详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "获取单本书籍",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "书籍 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据 ID 更新书籍信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "更新书籍",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "书籍 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据 ID 删除书籍",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "删除书籍",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "书籍 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    }
                }
            }
        },
...
        "/students/{id}/books": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定学生的所有借阅记录（包含书籍信息）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "borrow"
                ],
                "summary": "获取学生借书记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "学生 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Student"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    }
                }
            }
        },
        "/students/{student_id}/books/{book_id}/return": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "学生归还书籍",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "borrow"
                ],
                "summary": "归还书籍",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "学生 ID",
                        "name": "student_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "书籍 ID",
                        "name": "book_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Book_Student"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "middleware.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "models.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "book_students": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Book_Student"
                    }
                },
                "copies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BookCopy"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.BookCopy": {
            "type": "object",
            "properties": {
                "book_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.BookStatus"
                }
            }
        },
        "models.BookStatus": {
            "type": "string",
            "enum": [
                "available",
                "borrowed",
                "lost"
            ],
            "x-enum-comments": {
                "BookStatusAvailable": "在馆",
                "BookStatusBorrowed": "已借出",
                "BookStatusLost": "遗失"
            },
            "x-enum-descriptions": [
                "在馆",
                "已借出",
                "遗失"
            ],
            "x-enum-varnames": [
                "BookStatusAvailable",
                "BookStatusBorrowed",
                "BookStatusLost"
            ]
        },
        "models.Book_Student": {
            "type": "object",
            "properties": {
                "book_id": {
                    "type": "integer"
                },
                "borrowed_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "return_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.BorrowStatus"
                },
                "student_id": {
                    "type": "integer"
                }
            }
        },
        "models.BorrowStatus": {
            "type": "string",
            "enum": [
                "borrowed",
                "returned",
                "lost"
            ],
            "x-enum-varnames": [
                "BorrowStatusBorrowed",
                "BorrowStatusReturned",
                "BorrowStatusLost"
            ]
        },
        "models.Student": {
            "type": "object",
            "properties": {
                "book_student": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Book_Student"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}